+++ otlet/__init__.py
@@ -25,6 +25,7 @@ OR OTHER DEALINGS IN THE SOFTWARE.
 
 from .types import *
 from .http import *
+from .exceptions import *
 
-__version__ = '0.1.2'
+__version__ = '0.2.0'
 __license__ = 'MIT'